Move Driver interface to a new package
