+++ ecs-init/volumes/driver/interface.go
@@ -1,0 +1,35 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package driver
+
+import "github.com/aws/amazon-ecs-agent/ecs-init/volumes/types"
+
+// VolumeDriver contains the methods for volume drivers to implement
+type VolumeDriver interface {
+	Setup(string, *types.Volume)
+	Create(*CreateRequest) error
+	Remove(*RemoveRequest) error
+}
+
+// CreateRequest holds fields necessary for creating a volume
+type CreateRequest struct {
+	Name    string
+	Path    string
+	Options map[string]string
+}
+
+// RemoveRequest holds fields necessary for removing a volume
+type RemoveRequest struct {
+	Name string
+}
